requirements.txt
yfinance
pandas
scikit-learn
TA-Lib
matplotlib
numpy
tensorflow
feedparser 
openai
python-dotenv
